Added extra input variables in Find*Idx and MatchTrigObj functions and removed duplicate methods with _DQMOff suffix. Removed test lines which were causing additional loops. Changed MET to PuppiMET for Muon_MassId definition. Separated the DQMOff selections in helper_nano_dqmoff.py file. This file is called by performance_nano_dqmoff.py
